grib_io.c: restore signedness for arguments for seek
stdio.seek, which calls fseeko, uses "off_t" for argument, and
when seeking toward head of the file, passing negative off_t value
must be passed.

Currently when callling stdio.seek, "size_t" variable is passed
to the argument. On 64 bit system, the problem is hidden because
both size_t (unsigned) and off_t (signed) are 64 bit.
But on 32 bit system, as size_t is unsigned 32 bit and off_t is
signed 64 bit, when passing size_t "message_size - already_read"
(which is intended to be negative to seek toward head) is converted
to large unsigned 32 bit value, which is not intended.

The fix is to correctly cast to off_t.