tests/grib_lam_bf: use ieee instead of ieee64 for linux 32 bit
"unpackedSubsetPrecision" value is used for "ieee_floats".
ieee_floats=2 means bytes=8 and grib_long_to_ieee64 is used,
which is supported only when sizeof(double) == sizeof(long) .

Use unpackedSubsetPrecision=1 on linux 32 bit system, which means
grib_long_to_ieee is to be used.

Note that this test is currently disabled on Windows.