bitsPerValue test: restrict max value to 31 on linux 32 bit
On grib_bitsPerValue.sh, restrict maximum bitsPerValue value
(MAX_BPV) to 31 on linux 32 bit, higher value is not supported.

Note that currently on Windows system, MAX_BPV is set to 26 instead
of 31, but this should be able to be changed to 31 as well on
Windows as this should be fixed with
"decode_long_array: avoid overflow on 32 bit when overread" change.

Also, on linux 64 bit system, change MAX_BPV to 63. This should also
be fixed with the above change.